+++ src/main/java/view/ResultView.java
@@ -1,6 +1,6 @@
 package view;
 
-import domain.game.Results;
+import domain.result.Results;
 import domain.ladder.Bridge;
 import domain.ladder.Ladder;
 import domain.ladder.Line;
