+++ src/test/java/domain/game/GameResultTest.java
@@ -1,5 +1,6 @@
 package domain.game;
 
+import domain.result.Result;
 import domain.user.UserName;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
