+++ src/test/java/domain/game/LadderGameTest.java
@@ -1,6 +1,8 @@
 package domain.game;
 
 import domain.ladder.Ladder;
+import domain.result.Result;
+import domain.result.Results;
 import domain.user.Users;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
