+++ src/test/java/domain/game/ResultsTest.java
@@ -1,5 +1,7 @@
 package domain.game;
 
+import domain.result.Result;
+import domain.result.Results;
 import domain.user.Users;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
